refactor(backend): split PostgreSQL integration runner
